fix: load gcc-13-2-0 in jobscript template
Fix library errors in the cluster tests by loading gcc-13-2-0
from within the modules.

Set data type to "object" for the sample array of
the grid iterator. For now, this preserves the set data types of
the parameters, which enables e.g., to inject integers in the 4C
input files.